requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
decorator==5.2.1
defusedxml==0.8.0rc2
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==65.1.0
django-cors-headers==4.7.0
django-crispy-forms==1.6.0
django-registration-redux==1.4
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.5.0
google-auth==2.38.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
incremental==24.7.2
moviepy==2.1.2
msgpack==1.1.0
multidict==6.1.0
numpy==2.2.3
oauthlib==3.2.2
opencv-python==4.11.0.86
packaging==24.1
pillow==10.4.0
proglog==0.1.10
propcache==0.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyfcm==2.0.8
PyJWT==2.9.0
pymediainfo==7.0.1
pyOpenSSL==25.0.0
python-dotenv==1.0.1
python3-openid==3.2.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
service-identity==24.2.0
setuptools==75.8.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
stripe==11.2.0
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.18.3
zope.interface==7.2
